fix(inquiry): insertInquiry useGeneratedKeys 제거 + user /api/auth · /api/shared nginx 라우팅 (#13)
증상:
  1. POST /api/user/inquiries → 500 (MyBatis Jdbc3KeyGenerator NPE)
  2. POST https://academy.unmong.com/api/auth/signup → 401
     (nginx 가 /api/user/auth/signup 으로 잘못 rewrite)

원인:
  1. MyBatis useGeneratedKeys=true + keyProperty="csSeq" 가 @Param 기반 mapper
     메서드에 붙어있으면 parameter 를 Map/POJO 로 가정해 BeanWrapper NPE 발생.
     Service 에서 selectMyList(limit=1) 로 이미 재조회 중이라 생성 키 반환 불필요.
  2. nginx 의 user-web /api/ location 이 모든 /api/* 를 /api/user/* 로 rewrite.
     /api/auth/**, /api/shared/** 는 shared 엔드포인트라 rewrite 금지 (admin 쪽
     처럼 ^~ 블록 필요).

조치:
  1. inquiry.xml: insertInquiry 의 useGeneratedKeys/keyProperty 제거
  2. InquiryMapper.insertInquiry 반환 타입 long → int (affected rows)
  3. academy.conf 에 ^~ /api/auth/ 와 ^~ /api/shared/ 블록 추가 (user-web side)

관련: Phase D E2E 테스트 중 발견

+++ backend/src/main/java/com/academy/mapper/InquiryMapper.java
@@ -61,8 +61,8 @@ public interface InquiryMapper {
         @Param("rawOutput") String rawOutput
     );
 
-    /** 사용자 문의 신규 등록. generated cs_seq 반환 위해 VO 대신 Long 사용시 useGeneratedKeys 필요. */
-    long insertInquiry(
+    /** 사용자 문의 신규 등록. int = affected rows. Service 에서 selectMyList 로 최신 행 재조회. */
+    int insertInquiry(
         @Param("userId") String userId,
         @Param("name") String name,
         @Param("title") String title,
